test: move LoRARite optimizer test to test_optimizers.py
